requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bandit==1.8.3
click==8.1.8
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
rich==14.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
stevedore==5.4.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
